WIP: add tests
